requirements.txt
cryptography>=41.0.0
argon2-cffi>=21.3.0
customtkinter>=6.6.1